// src/app/main/main-body/common/custom-modal/modal-options.interface.ts
export interface ModalOptions {
  saveCallback?: Function
  hideCallback?: Function
  hideButton?: boolean
  hideHeader?: boolean
  hideFooter?: boolean
  modatTitle: string
  maxWidth?: number
  formType?: string
  hideCloseButton?: boolean,
  modalId?: string
  notchangemodaltitle?: boolean,
  notDisabledInViewMode?: boolean,
}